+++ src/workflow/__init__.py
@@ -6,5 +6,5 @@
 # Website: https://cv.songshgeo.com/
 
 """
-分析华南农业生计演变的多主体模型。
+分析模型结果的工作流
 """
